[Dashboard listing] update edit url and view url (#416)
* [Dashboard listing] update edit url and view url

Update to include BASE_URL if basePath is defined. Since create url is just
a direct nav and needs the basePath when passing props to the provider, the source
code no longer appends the basePath to keep these URLs consistent.

A feature should consolidate the basePaths in the dashboard listing service so that
createUrl also adds the base path. At which point we no longer need the base path.

Also removed some rogue URLs.

Core PR:
https://github.com/opensearch-project/OpenSearch-Dashboards/pull/3922

---------

Signed-off-by: Kawika Avilla <kavilla414@gmail.com>
Signed-off-by: Joshua Li <joshuali925@gmail.com>
Co-authored-by: Joshua Li <joshuali925@gmail.com>